refactor: view에서 사용할 날씨 정보를 모아둔 model의 이름을 entity로 변경하고, domain 레이어에 위치시켜 의존성 역전 원칙을 준수하도록 함
